chore: updated storybook
There seems to be a lot of breaking changes b/w storybook 5 and 8.
As a result, it seems like all component docs have to be rewritten.
I feel it would be better to focus of emotion upgrade first and then
relook at upgrading storybook. However, I am saving the changes for
future reference.